+++ ChangeLog.md
@@ -16,6 +16,7 @@
 	* Update 'Swapout check' script to use modules' modified flag to see if a module has modifications
 	* Update 'Commodity sale check' script to ensure that a commodity is in demand at the target station before reporting on a sale
   * Ensure that distance from home is updated whenever system co-ordinates are updated
+  * Ensure that speech responder's P() works for ships
 
 #2.1.0-b2
   * Add 'Engineer' government type for engineer bases